Move webserver from main file to seperate and add module imports
